+++ app/src/main/java/com/example/texteditor/MainActivity.kt
@@ -39,9 +39,8 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.canopas.editor.ui.RichTextEditor
-import com.canopas.editor.ui.RichTextStyle
-import com.canopas.editor.ui.RichTextValue
-import com.canopas.editor.ui.TextEditorValue
+import com.canopas.editor.ui.data.RichTextStyle
+import com.canopas.editor.ui.data.TextEditorValue
 import com.canopas.editor.ui.rememberEditorState
 import com.example.texteditor.ui.theme.TextEditorTheme
 
@@ -65,7 +64,6 @@ class MainActivity : ComponentActivity() {
 @Composable
 fun GreetingPreview() {
     TextEditorTheme {
-        var basicRichTextValue by remember { mutableStateOf(RichTextValue()) }
         var state by rememberEditorState()
 
         Column {
