+++ editor/src/main/java/com/canopas/editor/ui/data/RichTextPart.kt
@@ -1,4 +1,4 @@
-package com.canopas.editor.ui
+package com.canopas.editor.ui.data
 
 data class RichTextPart(
     val fromIndex: Int,
