+++ editor/src/main/java/com/canopas/editor/ui/model/RichTextValue.kt
@@ -1,0 +1,92 @@
+package com.canopas.editor.ui.model
+
+import androidx.compose.runtime.Immutable
+import androidx.compose.ui.text.SpanStyle
+import androidx.compose.ui.text.buildAnnotatedString
+import androidx.compose.ui.text.input.OffsetMapping
+import androidx.compose.ui.text.input.TextFieldValue
+import androidx.compose.ui.text.input.TransformedText
+import androidx.compose.ui.text.input.VisualTransformation
+import com.canopas.editor.ui.data.ContentType
+import com.canopas.editor.ui.data.ContentValue
+import com.canopas.editor.ui.data.RichTextPart
+import com.canopas.editor.ui.data.RichTextStyle
+import com.canopas.editor.ui.data.RichTextValueBuilder
+
+@Immutable
+internal data class RichTextValue internal constructor(
+    internal val textFieldValue: TextFieldValue,
+    internal val currentStyles: MutableSet<RichTextStyle> = mutableSetOf(),
+    internal val parts: MutableList<RichTextPart> = mutableListOf()
+) : ContentValue() {
+
+    constructor(
+        text: String = ""
+    ) : this(textFieldValue = TextFieldValue(text = text))
+
+    override val type: ContentType = ContentType.RICH_TEXT
+    override var isSelected: Boolean = false
+
+    internal val visualTransformation
+        get() = VisualTransformation {
+            TransformedText(
+                text = annotatedString,
+                offsetMapping = OffsetMapping.Identity
+            )
+        }
+
+    private val annotatedString
+        get() = buildAnnotatedString {
+            append(textFieldValue.text)
+            parts.map { part ->
+                val spanStyle = part.styles.fold(SpanStyle()) { spanStyle, richTextStyle ->
+                    richTextStyle.applyStyle(spanStyle)
+                }
+
+                addStyle(
+                    style = spanStyle,
+                    start = part.fromIndex,
+                    end = part.toIndex + 1,
+                )
+            }
+        }
+
+    fun toggleStyle(style: RichTextStyle): RichTextValue {
+        return if (currentStyles.contains(style)) {
+            removeStyle(style)
+        } else {
+            addStyle(style)
+        }
+    }
+
+    private fun addStyle(vararg style: RichTextStyle): RichTextValue {
+        return RichTextValueBuilder
+            .from(this)
+            .addStyle(*style)
+            .build()
+    }
+
+    private fun removeStyle(vararg style: RichTextStyle): RichTextValue {
+        return RichTextValueBuilder
+            .from(this)
+            .removeStyle(*style)
+            .build()
+    }
+
+    fun updateStyles(newStyles: Set<RichTextStyle>): RichTextValue {
+        return RichTextValueBuilder
+            .from(this)
+            .removeStyle(*this.currentStyles.toTypedArray())
+            .updateStyles(newStyles)
+            .build()
+    }
+
+    internal fun updateTextFieldValue(newTextFieldValue: TextFieldValue): RichTextValue {
+        return RichTextValueBuilder
+            .from(this)
+            .updateTextFieldValue(newTextFieldValue)
+            .build()
+    }
+
+    fun hasStyle(style: RichTextStyle) = currentStyles.contains(style)
+}
